builders: refactor using inheritance
