+++ inspire_schemas/builders/__init__.py
@@ -26,9 +26,9 @@
 
 from __future__ import absolute_import, division, print_function
 
+from .authors import AuthorBuilder  # noqa
+from .conferences import ConferenceBuilder  # noqa
+from .jobs import JobBuilder  # noqa
 from .literature import LiteratureBuilder  # noqa
 from .references import ReferenceBuilder  # noqa
 from .signatures import SignatureBuilder  # noqa
-from .authors import AuthorBuilder  # noqa
-from .jobs import JobBuilder  # noqa
-from .conferences import ConferenceBuilder  # noqa
